server/requirement.txt
django-cors-headers
django
PyJWT
djangorestframework